flask_site_checker/requirements.txt
Flask>=3.0.0
requests>=2.31.0
dnspython>=2.6.1
idna>=3.6
